tile->glyph.  For so many reasons
tile implies extra data *about* the tile.  stuff like its colour, position on the map, monster that's on top of it.  Glyph is just that, a glyph.  I want a glyph for the letter a, but that wouldn't make sense as a tile, since a is a monster that goes ON TOP OF a tile

Why are you reading this far down, you creep.

seriously, you've got to stop this habit